implement weld

+++ server-to-do-list/src/main/java/com/koror/app/endpoint/TaskEndpoint.java
@@ -1,43 +1,42 @@
 package com.koror.app.endpoint;
 
+import com.koror.app.api.service.ISessionService;
+import com.koror.app.api.service.ITaskService;
 import com.koror.app.entity.*;
 import com.koror.app.error.SessionNotValidateException;
-import com.koror.app.service.AssigneeTaskService;
-import com.koror.app.service.SessionService;
-import com.koror.app.service.TaskService;
+import org.jetbrains.annotations.Nullable;
 
+import javax.inject.Inject;
 import javax.jws.WebMethod;
 import javax.jws.WebParam;
 import javax.jws.WebService;
-import java.util.ArrayList;
 import java.util.List;
 
 @WebService
 public class TaskEndpoint {
 
-    private TaskService taskService;
+    @Inject
+    private ITaskService taskService;
 
-    private SessionService sessionService;
-
-    public TaskEndpoint(TaskService taskService, SessionService sessionService) {
-        this.taskService = taskService;
-        this.sessionService = sessionService;
-    }
+    @Inject
+    private ISessionService sessionService;
 
     @WebMethod
-    public Result addTask(@WebParam( name = "name", partName = "name")Task task, @WebParam( name = "session", partName = "session") Session session){
+    public Result addTask(
+            @WebParam(name = "name", partName = "name") @Nullable Task task,
+            @WebParam(name = "session", partName = "session") @Nullable Session session) {
         final boolean validateSession = sessionService.validate(session);
-        if(!validateSession) throw new SessionNotValidateException();
-        taskService.add(task,session.getUser());
+        if (session == null || !validateSession) throw new SessionNotValidateException();
+        taskService.add(task, session.getUser());
         final Result result = new Result();
         result.success();
         return result;
     }
 
     @WebMethod
-    public Result clearTask(@WebParam( name = "session", partName = "session") Session session){
+    public Result clearTask(@WebParam(name = "session", partName = "session") @Nullable Session session) {
         final boolean validateSession = sessionService.validate(session);
-        if(!validateSession) throw new SessionNotValidateException();
+        if (session == null || !validateSession) throw new SessionNotValidateException();
         final User user = session.getUser();
         final List<Task> taskList = taskService.getListTaskByUserId(user.getId());
         taskService.clearTask(taskList);
@@ -48,11 +47,11 @@ public class TaskEndpoint {
 
     @WebMethod
     public Result completeTask(
-            @WebParam( name = "id", partName = "id")String taskId,
-            @WebParam( name = "session", partName = "session") Session session){
-        final boolean validateSession =sessionService.validate(session);
-        if(!validateSession) throw new SessionNotValidateException();
-        final Task task =taskService.getById(taskId);
+            @WebParam(name = "id", partName = "id") @Nullable String taskId,
+            @WebParam(name = "session", partName = "session") @Nullable Session session) {
+        final boolean validateSession = sessionService.validate(session);
+        if (!validateSession) throw new SessionNotValidateException();
+        final Task task = taskService.getById(taskId);
         task.setComplete(true);
         taskService.completeTask(task);
         final Result result = new Result();
@@ -62,10 +61,10 @@ public class TaskEndpoint {
 
     @WebMethod
     public Result deleteTask(
-            @WebParam( name = "id", partName = "id")String taskId,
-            @WebParam( name = "session", partName = "session") Session session){
+            @WebParam(name = "id", partName = "id") @Nullable String taskId,
+            @WebParam(name = "session", partName = "session") @Nullable Session session) {
         final boolean validateSession = sessionService.validate(session);
-        if(!validateSession) throw new SessionNotValidateException();
+        if (session == null || !validateSession) throw new SessionNotValidateException();
         final User user = session.getUser();
         taskService.delete(taskId, user.getId());
         final Result result = new Result();
@@ -74,20 +73,20 @@ public class TaskEndpoint {
     }
 
     @WebMethod
-    public List<Task> getTaskList(@WebParam( name = "session", partName = "session") Session session) throws SessionNotValidateException{
-        final boolean validateSession =sessionService.validate(session);
-        if(!validateSession) throw new SessionNotValidateException();
+    public List<Task> getTaskList(@WebParam(name = "session", partName = "session") @Nullable Session session){
+        final boolean validateSession = sessionService.validate(session);
+        if (session == null || !validateSession) throw new SessionNotValidateException();
         User user = session.getUser();
         return taskService.getListTaskByUserId(user.getId());
     }
 
     @WebMethod
     public Result taskToGroupTask(
-            @WebParam( name = "task", partName = "task")Task task,
-            @WebParam( name = "group", partName = "group")Group group,
-            @WebParam( name = "session", partName = "session") Session session){
+            @WebParam(name = "task", partName = "task") @Nullable Task task,
+            @WebParam(name = "group", partName = "group") @Nullable Group group,
+            @WebParam(name = "session", partName = "session") @Nullable Session session) {
         final boolean validateSession = sessionService.validate(session);
-        if(!validateSession) throw new SessionNotValidateException();
+        if (!validateSession) throw new SessionNotValidateException();
         taskService.setGroup(task, group);
         final Result result = new Result();
         result.success();
@@ -96,10 +95,10 @@ public class TaskEndpoint {
 
     @WebMethod
     public Result updateTask(
-            @WebParam( name = "task", partName = "task")Task task,
-            @WebParam( name = "session", partName = "session") Session session){
-        final boolean validateSession =sessionService.validate(session);
-        if(!validateSession) throw new SessionNotValidateException();
+            @WebParam(name = "task", partName = "task") @Nullable Task task,
+            @WebParam(name = "session", partName = "session") @Nullable Session session) {
+        final boolean validateSession = sessionService.validate(session);
+        if (!validateSession) throw new SessionNotValidateException();
         taskService.update(task);
         final Result result = new Result();
         result.success();
